+++ fabric/src/main/java/immersive_aircraft/fabric/ClientFabric.java
@@ -6,6 +6,7 @@ import immersive_aircraft.client.KeyBindings;
 import immersive_aircraft.fabric.resources.FabricObjectLoader;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.minecraft.resource.ResourceType;
@@ -14,6 +15,8 @@ public final class ClientFabric implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         ClientLifecycleEvents.CLIENT_STARTED.register((event) -> ClientMain.postLoad());
+
+        ClientTickEvents.START_CLIENT_TICK.register((event) -> ClientMain.tick());
 
         ResourceManagerHelper.get(ResourceType.CLIENT_RESOURCES).registerReloadListener(new FabricObjectLoader());