+++ forge/src/main/java/immersive_aircraft/forge/ForgeBusEvents.java
@@ -18,5 +18,7 @@ public class ForgeBusEvents {
             ClientMain.postLoad();
             firstLoad = false;
         }
+
+        ClientMain.tick();
     }
 }