+++ changelog.md
@@ -9,6 +9,7 @@
 * You can no longer apply knockback to your vehicle
 * Fixed the Biplane phantom push when leaving
 * Fixed first person camera offset
+* Vehicles can have a separate camera perspective, and is third person by default (configurable)
 
 # 0.2.0